Migrated Document over to being rendered with a Visitor
